fix: rename to formFieldConstants
