+++ src/pages/EarthEngineImport/components/AggregationType.js
@@ -3,7 +3,10 @@ import { ReactFinalForm, SingleSelectFieldFF, hasValue } from '@dhis2/ui'
 import React from 'react'
 import { StyledField } from '../../../components/index.js'
 import { getEarthEngineConfigs } from '../util/earthEngines.js'
-import { AGGREGATION_TYPE, EARTH_ENGINE_ID } from '../util/getFormValues.js'
+import {
+    AGGREGATION_TYPE,
+    EARTH_ENGINE_ID,
+} from '../util/formFieldConstants.js'
 
 const AGGREGATION_TYPE_SUM = 'sum'
 const AGGREGATION_TYPE_MIN = 'min'
