+++ src/pages/EarthEngineImport/form-helper.js
@@ -1,36 +1,27 @@
 // import { FORM_ERROR, jobStartedMessage } from '../../utils/final-form.js'
 import { genericErrorMessage } from '../../utils/helper.js'
 import { extractIdAndMessage } from '../../utils/xhr.js'
-import {
-    PERIOD,
-    DATA_ELEMENT_ID,
-    BAND_COCS,
-    getFormValues,
-} from './util/getFormValues.js'
 
 const isAsync = true
 
 const onImport =
     ({ engine, setProgress, addTask, setShowFullSummaryTask }) =>
     async (values) => {
-        const { dryRun, eeData, ...rest } = values
-        const { dataElementId, period, ...bandCocs } = getFormValues(rest, [
-            PERIOD,
-            DATA_ELEMENT_ID,
-            BAND_COCS,
-        ])
+        const { dryRun, eeData, dataElementId, period, bandCocs } = values
 
         setProgress(true)
 
         let dataValues
 
-        if (Object.keys(bandCocs).length) {
+        if (bandCocs.length) {
             dataValues = eeData.map(({ ouId, bandId, value }) => {
                 return {
                     dataElement: dataElementId,
                     period,
                     orgUnit: ouId,
-                    categoryOptionCombo: bandCocs[bandId],
+                    categoryOptionCombo: bandCocs.find(
+                        (bc) => bc.bandId === bandId
+                    ),
                     value,
                 }
             })
