+++ src/pages/EarthEngineImport/util/formFieldConstants.js
@@ -1,0 +1,8 @@
+export const EARTH_ENGINE_ID = 'earthEngineId'
+export const PERIOD = 'period'
+export const ORGANISATION_UNITS = 'organisationUnits'
+export const ROUNDING = 'rounding'
+export const DATA_ELEMENT_ID = 'dataElementId'
+export const AGGREGATION_TYPE = 'aggregationType'
+export const BAND_COCS = 'bandCocs'
+export const ASSOCIATED_GEOMETRY = 'associatedGeometry'
